backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.29.0
httpx>=0.27.0
solana>=0.25.0
solders>=0.16.0
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
sqlalchemy>=2.0.25
alembic>=1.11.0
pydantic>=1.10.0
python-multipart>=0.0.5
psycopg2-binary>=2.9.0